% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiMaUmisc-package.R
\docType{package}
\name{BiMaUmisc-package}
\alias{BiMaUmisc}
\alias{BiMaUmisc-package}
\title{BiMaU Miscellaneous}
\description{
This package contains useful data analysis tools, such as a function to plot survival curves, a function to format \emph{p}-values, 
and a function (work in progress) to automatically select the test to apply when assessing differences in continuous variables between groups.
}
\section{Functions}{

\code{\link[BiMaUmisc]{show.p}}, \code{\link[BiMaUmisc]{survival.plot}}
}

\author{
\strong{Maintainer}: Anna Felip-Badia \email{annafelipibadia@gmail.com} (\href{https://orcid.org/0009-0002-8751-5228}{ORCID})

Authors:
\itemize{
  \item Anna Felip-Badia \email{annafelipibadia@gmail.com} (\href{https://orcid.org/0009-0002-8751-5228}{ORCID})
  \item Aleix Martín-Moral (\href{https://orcid.org/0009-0005-0529-3046}{ORCID})
  \item Sara Perez-Jaume (\href{https://orcid.org/0000-0002-9805-4810}{ORCID})
}

}
\keyword{internal}
