% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.plot.R
\name{survival.plot}
\alias{survival.plot}
\title{Plot survival curves}
\usage{
survival.plot(data, time, status, strata = 1,
              legend = TRUE, legend.pos = "bottomleft", legend.labs = NULL,
              legend.title = NULL, legend.cex = 1,
              events = FALSE, events.pos = NULL, events.text = NULL,
              mark.time = TRUE, lwd = 3, 
              risk = TRUE, risk.text = TRUE, risk.labs = NULL, risk.cex = 1, 
              p.value = TRUE, p.value.pos = NULL,
              units = "days", xlab = NULL, times = NULL,
              yscale = 1, ylab = NULL, 
              main = "Survival curve", col = NULL, ...)
}
\arguments{
\item{data}{name of the dataset, of class \code{data.frame}, containing at least 2 columns (follow up time, status indicator).}

\item{time}{follow-up time column name. Can be specified as a character string (with quotes) or as an unquoted symbol.}

\item{status}{status indicator column name. Can be specified as a character string (with quotes) or as an unquoted symbol.}

\item{strata}{strata indicator column name. Can be specified as a character string (with quotes) or as an unquoted symbol. Up to four categories. Must be a factor. \cr
The default value is 1, this is, there are no groups.}

\item{legend}{logical indicating whether a legend should be added to the plot. If \code{strata} is set to 1, this argument is ignored. \cr
The default value is \code{TRUE}.}

\item{legend.pos}{this argument can be a single keyword from the list in \code{\link[graphics]{legend}}, such as \code{"bottomleft"} or \code{"topright"},
which places the legend on the inside of the plot frame at the given location. In this case, this argument is passed to \code{\link[base]{match.arg}}. 
Otherwise, a 2-dimensional numerical vector indicating the x-y coordinates can be indicated. If \code{legend} is set to \code{FALSE}, this argument is ignored. \cr
The default value is \code{"bottomleft"}.}

\item{legend.labs}{a character vector of length \eqn{\geq}2 to appear as the legend labels. If \code{legend} is set to \code{FALSE}, this argument is ignored. \cr
The default value, \code{NULL}, indicates that the levels of \code{strata} are to be used as the legend labels.}

\item{legend.title}{title for the legend. If \code{legend} is set to \code{FALSE}, this argument is ignored. \cr
The default value, \code{NULL}, indicates on title should be added.}

\item{legend.cex}{legend's character expansion factor. A numerical value giving the amount by which the legend text should be magnified relative to the default.
If \code{legend} is set to \code{FALSE}, this argument is ignored. \cr
The default value is 1.}

\item{events}{logical indicating whether the number of events should be added to the text in the legend. 
If \code{strata} is set to 1, the total number of events is added directly to the plot instead.
If \code{legend} is set to \code{FALSE} and \code{strata} is not 1, this argument is ignored. \cr
The default value is \code{FALSE}.}

\item{events.pos}{2-dimensional numerical vector indicating the x-y coordinates where the number of events should be added to the plot.
If \code{strata} is not 1 or \code{events} is set to \code{FALSE}, this argument is ignored. \cr
The default value, \code{NULL}, indicates that the text should be placed on the bottom-left corner.}

\item{events.text}{a character indicating the text to display before the number of events in the plot.
If \code{events} is set to \code{FALSE}, this argument is ignored. \cr
The default value, \code{NULL}, indicates that the text "Events" should be added when \code{strata} is set to 1 and the text "events" should be added when \code{strata} is not set to 1.}

\item{mark.time}{logical indicating whether the curves should be marked at each censoring time. \cr
The default value is \code{TRUE}.}

\item{lwd}{a number indicating the line width(s). Alternatively, when \code{strata} is not set to 1, 
an \eqn{n}-dimensional numerical vector can be indicated, where \eqn{n} is the number of levels in \code{strata}. \cr
The default value is 3.}

\item{risk}{logical indicating whether the number-at-risk table should be appended to the plot. \cr
The default value is \code{TRUE}.}

\item{risk.text}{logical indicating whether the legend for the number-at-risk table should display text labels. When \code{FALSE}, line segments are used instead.
If \code{risk} is set to \code{FALSE} or \code{strata} is set to 1, this argument is ignored. \cr
The default value is \code{TRUE}.}

\item{risk.labs}{\eqn{n}-dimensional character vector, where \eqn{n} is the number of levels in \code{strata}, indicating the text labels in the number-at-risk table.
If \code{risk.text} is set to \code{FALSE} or \code{strata} is set to 1, this argument is ignored. \cr
The default value, \code{NULL}, indicates that the levels of \code{strata} should be used.}

\item{risk.cex}{number-at-risk table's character expansion factor. A numerical value giving the amount by which the number-at-risk table text should be magnified relative to the default.
If \code{risk} is set to \code{FALSE}, this argument is ignored. \cr
The default value is 1.}

\item{p.value}{logical indicating whether the \emph{p}-value obtained from a call to \code{\link[survival]{survdiff}} with \code{rho} equal to 0, this is the log-rank or Mantel-Haenszel test, 
should be displayed in the plot. The \emph{p}-value will be displayed in the format specified by the \code{\link[BiMaUmisc]{show.p}} function.
If \code{strata} is set to 1, this argument is ignored. \cr
The default value is \code{TRUE}.}

\item{p.value.pos}{2-dimensional numerical vector indicating the x-y coordinates where the number of events should be added to the plot.
If \code{strata} is not 1 or \code{events} is set to \code{FALSE}, this argument is ignored. \cr
The default value, \code{NULL}, indicates that the text should be placed on the upper-right corner.}

\item{units}{character indicating the time units. Options are: "days", "weeks", "months" or "years".
This argument will be used to create the default label and tick-marks for the x-axis. This argument is not case-sensitive and is passed to \code{\link[base]{match.arg}}. \cr
The default value is \code{"days"}.}

\item{xlab}{character indicating a label for the x-axis. \cr
The default value, \code{NULL}, indicates that indicates that the text "Time (units)" should be used, 
where "units" is specified by the \code{units} argument.}

\item{times}{numerical vector with the time points at which tick-marks are to be drawn in the x-axis. \cr
The default value, \code{NULL}, indicates that a time vector will be automatically constructed based on the \code{units} argument:
\itemize{
  \item If \code{units} is set to "days": from 0 to the maximum survival time rounded up to the nearest multiple of 7.
  \item If \code{units} is set to "weeks": from 0 to the maximum survival time rounded up to the nearest integer.
  \item If \code{units} is set to "months": from 0 to the maximum survival time rounded up to the nearest multiple of 12.
  \item If \code{units} is set to "years": from 0 to the maximum survival time rounded up to the nearest integer.
}}

\item{yscale}{a numeric value used to multiply the labels on the y-axis. A value of 100, for instance, would be used to give a percent scale.
Notice only the labels are changed, not the actual plot coordinates. \cr
The default value is 1.}

\item{ylab}{character indicating a label for the y-axis. \cr
The default value, \code{NULL}, indicates that indicates that the text "Survival" should be used. 
Moreover, if \code{yscale} is set to 1 (or 100) the text "(proportion)" or "(percentage)" will also be added.}

\item{main}{overall title for the plot. \cr
The default value is \code{"Survival curve"}.}

\item{col}{colour palette to be used. \cr
The default value, \code{NULL}, indicates that a different colour should be used per level in \code{strata}. 
This is, a vector from 1 to \eqn{n}, where \eqn{n} is the number of levels in \code{strata}, will be used.}

\item{...}{other graphical parameters (to be passed to \code{\link[survival]{plot.survfit}}).}
}
\value{
A customized survival curve plot displaying the requested graphical and statistical modifications.
}
\description{
This function produces an elegant survival curve plot. 
It can automatically append the number-at-risk table and display the log-tank test \emph{p}-value (obtained from a call to \code{\link[survival]{survdiff}}). 
Highly customizable, it offers numerous arguments to easily modify the plot's appearance.
}
\section{Note}{

A call to \code{\link[graphics]{par}} is used in this function. Notice that the arguments
\code{font.axis}, \code{font.lab}, \code{cex.lab}, \code{las} and \code{xpd} are always set to 2, 2, 1.2, 1, and \code{TRUE}, respectively. Moreover,
the argument \code{mar} is always modified and depends on the number of levels in strata as well as whether or not the number-at-risk table is appended. \cr
For optimal resolution and layout alignment when using the function in an R file, adjust the plot window. 
For example, ideal dimensions are approximately 7 x 5'' for 24'' screens and 5.9 x 4.2'' for 13'' screens. \cr
For optimal resolution and layout alignment when using the function in an R markdown file or similar, use \code{fig.dim = c(7, 5)} in the corresponding chunk.
}

\examples{
lung <- survival::cancer
lung$time_y <- lung$time/365.25
lung$sex <- factor(lung$sex, levels = 1:2, labels = c("Male", "Female"))

# no strata
survival.plot(lung, time_y, status, units = "y")
# customizing
survival.plot(lung, time_y, status, units = "y",
              xlab = "Time since diagnosis (in years)",
              yscale = 100,
              main = "Overall survival", mark.col = "darkgray")

# 2 strata
survival.plot(lung, time_y, status, sex, units = "y")
# customizing
survival.plot(lung, time_y, status, sex, 
              legend.pos = "topright", legend.title = "Sex", 
              events = TRUE, risk.text = FALSE, p.value.pos = c(2.4, 0.4),
              units = "y", xlab = "Time since diagnosis (in years)", 
              main = "Overall survival", col = c("darkviolet", "darkgreen"))

}
